:recycle: Refactor Model package to Entity
